# openapi-spec-formatted.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Smart Healthcare System API",
        "version": "1.0.0",
        "description": "API documentation for Smart Healthcare System - UTS IAE/Web Service Development",
        "contact": {
            "name": "Healthcare Team"
        }
    },
    "servers": [
        {
            "url": "http://localhost:3000",
            "description": "Development server (API Gateway)"
        }
    ],
...
    "components": {
        "schemas": {
...
            "Error": {
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string",
                        "example": "Error message"
                    },
                    "error": {
                        "type": "string",
                        "example": "Detailed error description"
                    }
                }
            },
            "Success": {
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": true
                    },
                    "message": {
                        "type": "string",
                        "example": "Operation successful"
                    },
                    "data": {
                        "type": "object"
                    }
                }
            }
        }
    },
    "paths": {
        "/api/patients": {
            "get": {
                "summary": "Get all patients",
                "tags": [
                    "Patients"
                ],
                "responses": {
                    "200": {
                        "description": "List of all patients",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "count": {
                                            "type": "integer"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Patient"
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "summary": "Create new patient",
                "tags": [
                    "Patients"
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "required": [
                                    "name",
                                    "birth_date",
                                    "gender",
                                    "phone"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "birth_date": {
                                        "type": "string",
                                        "example": "1990-05-15"
                                    },
                                    "gender": {
                                        "type": "string",
                                        "enum": [
                                            "male",
                                            "female"
                                        ],
                                        "example": "male"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "081234567890"
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "Jl. Merdeka No. 123, Jakarta"
                                    },
                                    "blood_type": {
                                        "type": "string",
                                        "enum": [
                                            "O+",
                                            "A+",
                                            "B+",
                                            "AB+",
                                            "O-",
                                            "A-",
                                            "B-",
                                            "AB-"
                                        ],
                                        "example": "A+"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Patient created successfully"
                    },
                    "400": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/patients/{id}": {
            "get": {
                "summary": "Get patient by ID",
                "tags": [
                    "Patients"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "description": "Patient ID"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Patient details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Success"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Patient not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "summary": "Update patient",
                "tags": [
                    "Patients"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "description": "Patient ID"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "birth_date": {
                                        "type": "string"
                                    },
                                    "gender": {
                                        "type": "string"
                                    },
                                    "phone": {
                                        "type": "string"
                                    },
                                    "address": {
                                        "type": "string"
                                    },
                                    "blood_type": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Patient updated successfully"
                    },
                    "404": {
                        "description": "Patient not found"
                    }
                }
            },
            "delete": {
                "summary": "Delete patient",
                "tags": [
                    "Patients"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "description": "Patient ID"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Patient deleted successfully"
                    },
                    "404": {
                        "description": "Patient not found"
                    }
                }
            }
        },
        "/api/doctors": {
            "get": {
                "summary": "Get all doctors",
                "tags": [
                    "Doctors"
                ],
                "responses": {
                    "200": {
                        "description": "List of all doctors"
                    }
                }
            },
            "post": {
                "summary": "Create new doctor",
                "tags": [
                    "Doctors"
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "required": [
                                    "name",
                                    "specialization",
                                    "phone"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Dr. Jane Smith, Sp.PD"
                                    },
                                    "specialization": {
                                        "type": "string",
                                        "example": "Cardiologist"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "081234567891"
                                    },
                                    "schedule": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        },
                                        "example": [
                                            "Monday 09:00-12:00",
                                            "Wednesday 14:00-17:00"
                                        ]
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Doctor created successfully"
                    }
                }
            }
        },
...
        "/api/appointments/{id}": {
            "get": {
                "summary": "Get appointment by ID",
                "tags": [
                    "Appointments"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Appointment details"
                    }
                }
            },
            "delete": {
                "summary": "Delete appointment",
                "tags": [
                    "Appointments"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Appointment deleted successfully"
                    }
                }
            }
        },
...
        "/api/records/{id}": {
            "get": {
                "summary": "Get medical record by ID",
                "tags": [
                    "Medical Records"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Medical record details"
                    }
                }
            }
        },
        "/api/records/patient/{patientId}": {
            "get": {
                "summary": "Get medical records by patient ID",
                "tags": [
                    "Medical Records"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "patientId",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "description": "Patient ID"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of patient's medical records"
                    }
                }
            }
        }
    }
}